+++ apps/app/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @fodmap/app
 
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @fodmap/ui@1.0.2
+
 ## 0.1.4
 
 ### Patch Changes